requirements.txt
httpx # tg_raw_sender
colorlog # logging.helpers
python-dotenv # settings.config